update the endDate on startDate change (#7052)
