+++ src/domain/timeline/calendar/views/CalendarEventForm.tsx
@@ -1,36 +1,24 @@
 import React, { ReactNode, useMemo } from 'react';
-import { Form, Formik } from 'formik';
+import { Formik } from 'formik';
 import * as yup from 'yup';
 import { useTranslation } from 'react-i18next';
-import { Box, DialogContent, FormControlLabel, Switch } from '@mui/material';
 import DialogHeader, { DialogHeaderProps } from '../../../../core/ui/dialog/DialogHeader';
 import { BlockTitle } from '../../../../core/ui/typography';
-import FormikDatePicker from '../../../../core/ui/forms/DatePicker/FormikDatePicker';
 import { CalendarEventFormData } from '../CalendarEventsContainer';
 import { CalendarEventType } from '../../../../core/apollo/generated/graphql-schema';
-import Gutters from '../../../../core/ui/grid/Gutters';
-import FormikInputField from '../../../../core/ui/forms/FormikInputField/FormikInputField';
-import { gutters } from '../../../../core/ui/grid/utils';
-import FormikTimePicker from '../../../../core/ui/forms/DatePicker/FormikTimePicker';
-import FormikAutocomplete, { FormikSelectValue } from '../../../../core/ui/forms/FormikAutocomplete';
-import { Actions } from '../../../../core/ui/actions/Actions';
-import FormikMarkdownField from '../../../../core/ui/forms/MarkdownInput/FormikMarkdownField';
-import { TagsetField } from '../../../platform/admin/components/Common/TagsetSegment';
-import GridProvider from '../../../../core/ui/grid/GridProvider';
 import { displayNameValidator } from '../../../../core/ui/forms/validator';
 import { CalendarEventDetailData } from '../CalendarEventDetailContainer';
-import { LoadingButton } from '@mui/lab';
 import { MARKDOWN_TEXT_LENGTH } from '../../../../core/ui/forms/field-length.constants';
 import MarkdownValidator from '../../../../core/ui/forms/MarkdownInput/MarkdownValidator';
 import dayjs from 'dayjs';
-import FormikDurationMinutes from '../../../../core/ui/forms/DatePicker/FormikDurationMinutes';
 import { isSameDay } from '../../../../core/utils/time/utils';
+import EventForm from './EventForm';
+import { FormikSelectValue } from '../../../../core/ui/forms/FormikSelect';
+import GridProvider from '../../../../core/ui/grid/GridProvider';
 
 const DEFAULT_DURATION_MINUTES = 30;
 
-type DateType = Date | undefined;
-
-interface CalendarEventFormProps {
+export interface CalendarEventFormProps {
   event: Partial<CalendarEventDetailData> | undefined;
   dialogTitle: string;
   onClose: DialogHeaderProps['onClose'];
@@ -155,14 +143,6 @@ const CalendarEventForm = ({
     }),
   });
 
-  const getMinTime = (startDate: DateType, endDate: number | DateType) => {
-    if (!startDate) {
-      return undefined;
-    }
-
-    return isSameDay(startDate, endDate) ? undefined : dayjs(startDate);
-  };
-
   return (
     <GridProvider columns={12}>
       <DialogHeader onClose={onClose}>
@@ -174,97 +154,7 @@ const CalendarEventForm = ({
         validationSchema={validationSchema}
         enableReinitialize
       >
-        {({ isValid, handleSubmit, values: { wholeDay, startDate, endDate }, setFieldValue }) => (
-          <>
-            <DialogContent>
-              <Form>
-                <Gutters disablePadding>
-                  <Gutters disablePadding sx={{ flexDirection: { xs: 'column', sm: 'row' } }}>
-                    <Box sx={{ flex: 1 }}>
-                      <FormikInputField name="displayName" title={t('fields.displayName')} />
-                    </Box>
-                    <FormikAutocomplete
-                      name="type"
-                      label={t('calendar.event.type')}
-                      values={typeOptions}
-                      sx={{ flex: 1 }}
-                    />
-                  </Gutters>
-                  <Gutters disablePadding sx={{ flexDirection: { xs: 'column', sm: 'row' } }}>
-                    <Gutters disablePadding sx={{ flexDirection: 'row', flexGrow: 1 }}>
-                      <FormikDatePicker name="startDate" label={t('common.date')} minDate={new Date()} />
-                      <FormikTimePicker
-                        name="startDate"
-                        label={t('fields.startTime')}
-                        containerProps={{ flexGrow: 1 }}
-                        disabled={wholeDay}
-                      />
-                      <FormControlLabel
-                        control={
-                          <Switch
-                            checked={wholeDay}
-                            name="wholeDay"
-                            onChange={() => {
-                              setFieldValue('wholeDay', !wholeDay);
-                            }}
-                          />
-                        }
-                        label={t('calendar.event.whole-day')}
-                        sx={{ flexShrink: 0 }}
-                      />
-                    </Gutters>
-                    <Gutters disablePadding sx={{ flexDirection: 'row', flexGrow: 1 }}>
-                      <FormikDatePicker
-                        name="endDate"
-                        label={t('common.date')}
-                        minDate={startDate}
-                        disabled={wholeDay}
-                      />
-                      {isSameDay(startDate, endDate) ? (
-                        <FormikDurationMinutes
-                          name="durationMinutes"
-                          startTimeFieldName="startDate"
-                          label={t('fields.endTime')}
-                          containerProps={{ flexGrow: 1 }}
-                          disabled={wholeDay}
-                        />
-                      ) : (
-                        <FormikTimePicker
-                          name="endDate"
-                          label={'End Time'}
-                          containerProps={{ flexGrow: 1 }}
-                          disabled={wholeDay}
-                          minTime={getMinTime(startDate, endDate)}
-                        />
-                      )}
-                    </Gutters>
-                  </Gutters>
-                  <FormikMarkdownField
-                    name="description"
-                    title={t('common.description')}
-                    maxLength={MARKDOWN_TEXT_LENGTH}
-                    sx={{ marginBottom: gutters(-1) }}
-                  />
-                  <Gutters disablePadding sx={{ flexDirection: 'row', flexGrow: 1 }}>
-                    <FormikInputField name="location.city" title={'Location'} placeholder={' '} fullWidth />
-                    <TagsetField name="tags" title={t('common.tags')} />
-                  </Gutters>
-                </Gutters>
-              </Form>
-            </DialogContent>
-            <Actions justifyContent="space-between" padding={gutters()}>
-              {actions}
-              <LoadingButton
-                variant="contained"
-                disabled={!isValid}
-                loading={isSubmitting}
-                onClick={() => handleSubmit()}
-              >
-                {t('buttons.save')}
-              </LoadingButton>
-            </Actions>
-          </>
-        )}
+        <EventForm typeOptions={typeOptions} isSubmitting={isSubmitting} actions={actions} />
       </Formik>
     </GridProvider>
   );
